backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
openai==1.3.0
google-generativeai==0.3.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
duckduckgo-search==4.1.1

